services/file_selection_service/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
drf-pydantic==1.0.0
pydantic==1.10.13
loguru==0.7.2
gitpython==3.1.40